+++ src/main/java/walkingkooka/spreadsheet/server/formatter/SpreadsheetFormatterMenu.java
@@ -20,10 +20,10 @@ package walkingkooka.spreadsheet.server.formatter;
 import walkingkooka.collect.list.Lists;
 import walkingkooka.plugin.PluginSelectorMenu;
 import walkingkooka.plugin.PluginSelectorMenuLike;
-import walkingkooka.spreadsheet.format.SpreadsheetFormatterInfo;
-import walkingkooka.spreadsheet.format.SpreadsheetFormatterName;
-import walkingkooka.spreadsheet.format.SpreadsheetFormatterProvider;
-import walkingkooka.spreadsheet.format.SpreadsheetFormatterSelector;
+import walkingkooka.spreadsheet.format.provider.SpreadsheetFormatterInfo;
+import walkingkooka.spreadsheet.format.provider.SpreadsheetFormatterName;
+import walkingkooka.spreadsheet.format.provider.SpreadsheetFormatterProvider;
+import walkingkooka.spreadsheet.format.provider.SpreadsheetFormatterSelector;
 import walkingkooka.text.printer.IndentingPrinter;
 import walkingkooka.tree.json.JsonNode;
 import walkingkooka.tree.json.marshall.JsonNodeContext;
