+++ src/main/java/walkingkooka/spreadsheet/server/parser/SpreadsheetParserSelectorEditContexts.java
@@ -20,7 +20,7 @@ package walkingkooka.spreadsheet.server.parser;
 import walkingkooka.plugin.ProviderContext;
 import walkingkooka.reflect.PublicStaticHelper;
 import walkingkooka.spreadsheet.format.SpreadsheetFormatterContext;
-import walkingkooka.spreadsheet.format.SpreadsheetFormatterProvider;
+import walkingkooka.spreadsheet.format.provider.SpreadsheetFormatterProvider;
 import walkingkooka.spreadsheet.parser.SpreadsheetParserContext;
 import walkingkooka.spreadsheet.parser.provider.SpreadsheetParserProvider;
 public final class SpreadsheetParserSelectorEditContexts implements PublicStaticHelper {
